requirements.txt
uvicorn==0.13.4
fastapi==0.63.0
python-escpos==2.2.0
dataset==1.4.5
arrow==1.0.2
black==20.8b1
